docs: Extract GitHub edit link to separate component
Refactors the GitHub edit link functionality into a dedicated `EditOnGithub` component, simplifying the page component and improving modularity. Removes inline link generation logic in favor of a more reusable approach.
