+++ apps/docs/app/docs/[[...slug]]/page.tsx
@@ -5,30 +5,10 @@ import {
 	DocsPage,
 	DocsTitle,
 } from "fumadocs-ui/page";
-import { SquarePen } from "lucide-react";
-import Link from "next/link";
 import { notFound } from "next/navigation";
+import { EditOnGithub } from "~/components/page/edit-on-github";
 import { Separator } from "~/components/ui/separator";
 import { source } from "~/lib/source";
-
-const getLinkTitleAndHref = (path?: string) => {
-	try {
-		if (!path) throw new Error("Path is required");
-		const githubUrl = process.env.NEXT_PUBLIC_GITHUB_URL;
-		if (!githubUrl) throw new Error("NEXT_PUBLIC_GITHUB_URL is not configured");
-		const defaultBranch = process.env.NEXT_PUBLIC_GITHUB_BRANCH ?? "main";
-		const cleanUrl = githubUrl.replace(/\/$/, "");
-		// Use URL API for robust parsing
-		const url = new URL(cleanUrl);
-		const [owner, repo] = url.pathname.split("/").filter(Boolean).slice(-2);
-		if (!owner || !repo) throw new Error("Invalid GitHub URL format");
-		const title = `Editing ${repo}/${path} at ${defaultBranch} · ${owner}/${repo}`;
-		const href = `${githubUrl}/edit/${defaultBranch}/${path}`;
-		return { title, href };
-	} catch {
-		return { title: "Edit this page on GitHub", href: "#" };
-	}
-};
 
 export default async function Page(props: {
 	params: Promise<{ slug?: string[] }>;
@@ -50,15 +30,7 @@ export default async function Page(props: {
 					</DocsBody>
 				</div>
 				<div className="flex flex-col pt-16">
-					<Link
-						{...getLinkTitleAndHref(`apps/docs/content/docs/${page.file.path}`)}
-						target="_blank"
-						rel="noopener noreferrer"
-						className="flex items-center gap-2 text-sm text-gray-600 hover:text-gray-900 hover:underline transition-colors"
-					>
-						<SquarePen className="w-4 h-4" />
-						Edit this page on GitHub
-					</Link>
+					<EditOnGithub path={page.file.path} />
 					<div className="mt-8">
 						<Separator />
 					</div>
