+++ apps/docs/components/page/edit-on-github.tsx
@@ -1,0 +1,48 @@
+
+import Link from "next/link";
+import { SquarePen } from "lucide-react";
+
+const getLinkTitleAndHref = (path?: string) => {
+	try {
+		if (!path) throw new Error("Path is required");
+		const githubUrl = process.env.NEXT_PUBLIC_GITHUB_URL;
+		if (!githubUrl) throw new Error("NEXT_PUBLIC_GITHUB_URL is not configured");
+		const defaultBranch = process.env.NEXT_PUBLIC_GITHUB_BRANCH ?? "main";
+		const cleanUrl = githubUrl.replace(/\/$/, "");
+		// Use URL API for robust parsing
+		const url = new URL(cleanUrl);
+		const [owner, repo] = url.pathname.split("/").filter(Boolean).slice(-2);
+		if (!owner || !repo) throw new Error("Invalid GitHub URL format");
+		const title = `Editing ${repo}/${path} at ${defaultBranch} · ${owner}/${repo}`;
+		const href = `${githubUrl}/edit/${defaultBranch}/${path}`;
+		return { title, href };
+	} catch {
+		return { title: "Edit this page on GitHub", href: "#" };
+	}
+};
+
+type EditOnGithubProps = {
+	/**
+	 * The path to the file to edit on GitHub.
+	 * 
+	 * @example
+	 * ```ts
+	 * <EditOnGithub path="/index.mdx" />
+	 * ```
+	 */
+  path: string;
+}
+
+export const EditOnGithub = ({ path }: EditOnGithubProps) => {
+  return (
+    <Link
+      {...getLinkTitleAndHref(`apps/docs/content/docs/${path}`)}
+      target="_blank"
+      rel="noopener noreferrer"
+    className="flex items-center gap-2 text-sm text-gray-600 hover:text-gray-900 hover:underline transition-colors"
+  >
+			<SquarePen className="w-4 h-4" />
+			Edit this page on GitHub
+		</Link>
+	);
+};
